refactor(content): shrink content pack index responsibilities
